+++ zwave-js/events.js
@@ -1,0 +1,8 @@
+const { EventEmitter } = require('events');
+
+const EE = new EventEmitter();
+EE.setMaxListeners(0);
+
+module.exports = {
+	NodeEventEmitter: EE
+};